requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.4.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
filelock==3.6.0
flatbuffers==2.0
frozenlist==1.2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
numexpr==2.8.3
numpy==1.23.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandas-datareader==0.10.0
Pillow==9.2.0
pip==22.1.2
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.43
scikit-learn==1.0.2
scipy==1.9.1
setuptools==63.4.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.11.4
torch==1.10.2
torchmetrics==0.10.0
torchvision==0.11.0a0
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2021.1
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.0.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.0
